Frkaya/gccfix07102019 (dotnet/coreclr#25624)
* Remove duplicate definition

* Fix conversion error

* 1ui64 doesn't exist on GCC


Commit migrated from https://github.com/dotnet/coreclr/commit/41d0619707a29b910b7d8056a3b5d4423fc928d2

+++ src/clr/src/ToolBox/superpmi/superpmi-shared/methodcontext.cpp
@@ -4366,7 +4366,7 @@ void MethodContext::recGetFieldType(CORINFO_FIELD_HANDLE  field,
         // class handle (we use only 'field' and 'memberParent' as keys).
         // Update the value in that case.
         unsigned index = GetFieldType->GetIndex(key);
-        if ((index != -1) && (GetFieldType->GetItem(index).A == 0))
+        if ((index != (unsigned)-1) && (GetFieldType->GetItem(index).A == 0))
         {
             GetFieldType->Update(index, value);
             DEBUG_REC(dmpGetFieldType(key, value));
